Allow to CRUD public metadata for STAFF by APPs
