+++ saleor/graphql/meta/mutations/delete_private_metadata.py
@@ -1,5 +1,7 @@
 import graphene
 
+from ....core.exceptions import PermissionDenied
+from ....permission.enums import AccountPermissions
 from ...core import ResolveInfo
 from ...core.types import MetadataError, NonNullList
 from ..permissions import PRIVATE_META_PERMISSION_MAP
@@ -27,6 +29,15 @@ class DeletePrivateMetadata(BaseMetadataMutation):
             description="Metadata keys to delete.",
             required=True,
         )
+
+    @classmethod
+    def check_permissions(cls, context, permissions=None, **data):  # type: ignore[override]
+        is_app = bool(getattr(context, "app", None))
+        if is_app and permissions and AccountPermissions.MANAGE_STAFF in permissions:
+            raise PermissionDenied(
+                message="Apps are not allowed to perform this mutation."
+            )
+        return super().check_permissions(context, permissions)
 
     @classmethod
     def perform_mutation(  # type: ignore[override]
